% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{chronoloG}
\alias{chronoloG}
\title{chronoloG}
\usage{
chronoloG(
  n_samples,
  n_samples_is_average = T,
  starting_cells = 1,
  s = 0.004,
  k = 3,
  max_gens = 10000,
  max_cells = 1e+12,
  n_markers = 58,
  mu = 1e-04,
  beta_params = c(2, 2)
)
}
\description{
chronoloG
}
