% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{get_genotypes_with_cnas}
\alias{get_genotypes_with_cnas}
\title{get_genotypes_with_cnas}
\usage{
get_genotypes_with_cnas(
  tree,
  n_markers,
  mu.indel,
  mu.cna,
  gens_until_first_cancer_cell
)
}
\description{
get_genotypes_with_cnas
}
