% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{get_purity}
\alias{get_purity}
\title{get_purities}
\usage{
get_purity(
  tree,
  purity_shape1 = 2,
  purity_shape2 = 2,
  clonality_min = 0.5,
  clonality_max = 1,
  sim_clonality = F
)
}
\description{
get_purities
}
