% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{get_purity_and_clonality}
\alias{get_purity_and_clonality}
\title{get_purity_and_clonality}
\usage{
get_purity_and_clonality(
  tree,
  purity_shape1 = 2,
  purity_shape2 = 2,
  clonality_min = 0.5,
  clonality_max = 0.95
)
}
\description{
get_purity_and_clonality
}
