% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_simulated_tree}
\alias{plot_simulated_tree}
\title{plot_simulated_tree}
\usage{
plot_simulated_tree(
  tree,
  layout = "ape",
  title = NA,
  purities = NULL,
  legend.position = "right"
)
}
\description{
plot_simulated_tree
}
