% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{random_generations}
\alias{random_generations}
\title{random_generations}
\usage{
random_generations(
  starting_cells = 1,
  s = 0.004,
  k = 3,
  max_gens = 10000,
  max_cells = 1e+12
)
}
\description{
random_generations
}
